improve antonkanerva's solution
