+++ challenges/06-maybe-keyword/solutions/kanerva/maybelib.py
@@ -1,4 +1,6 @@
+import __main__
 from random import choice
+from threading import Thread
 
 class MaybeBoolean():
 	def __repr__(self):
@@ -11,5 +13,11 @@ def is_instance(obj1, obj2):
 	if isinstance(obj1, MaybeBoolean):
 		return True
 	return isinstance(obj1, obj2)
+
+class MaybeSyntaxErrorListener(Thread):
+	def run(self):
+		while True:
+			if not isinstance(__main__.Maybe, MaybeBoolean):
+				raise SyntaxError("can't assign to keyword!")
 
 Maybe = MaybeBoolean()
